// src/api/types/models.d.ts
export interface BindVolunteerInformation {
  // 身份证号
  IDCard?: string;
  // 姓名
  name?: string;
}
// ...
export interface VolunteerInformation {
  // 详细地址
  address?: string;
  // 城市
  city?: string;
  // 国家
  country?: string;
  // 区
  district?: string;
  gender?: string;
  // ID
  id?: number;
  idcard?: string;
  // 姓名
  name?: string;
  // 当前页
  pageNo?: number;
  // 每页条数
  pageSize?: number;
  // 省份
  province?: string;
  // 注册时间
  registerTime?: string;
  // 性别
  sex?: number;
  volunteer?: Volunteer;
  volunteerId?: number;
}

export interface ResponseData<T> {
  // 返回数据
  data?: T;
  // 状态信息
  message?: string;
  // refreshToken
  refreshToken?: string;
  // 状态码
  status?: number;
  // accessToken
  token?: string;
}

export type JavaList<T> = Array<T>;

export interface Volunteer {
  // 头像
  avatarUrl?: string;
  // 志愿者ID
  id?: number;
  // 位置所在纬度
  latitude?: number;
  // 位置所在经度
  longitude?: number;
  // 昵称
  nickName?: string;
  // 是否在线
  online?: number;
  // 手机号
  phone?: string;
  // 用户状态
  state?: number;
  userinfo?: object;
  // 志愿者信息
  volunteerInformation?: VolunteerInformation;
}
// ...
export interface VolunteerAvatarUrl {
  // 头像链接
  avatarUrl?: string;
}

export interface VolunteerBindPhone {
  // 验证码
  code?: string;
  // 手机号
  phone?: string;
}

export interface VolunteerCase {
  // 案件ID
  caseId?: number;
  // 装备是否齐全，1为齐全，2为不齐全
  equipment?: number;
  // 关联表ID
  id?: number;
  // 状态（1为接受，2为拒绝，3为退出）
  state?: number;
  // 交通工具
  traffic?: string;
  // 志愿者
  volunteer?: Volunteer;
  // 志愿者ID
  volunteerId?: number;
}

export interface VolunteerLoginCode {
  // 从微信获取的code
  code?: string;
}

export interface VolunteerRefreshToken {
  // refreshToken
  refreshToken?: string;
}

export interface VolunteerUnbindCode {
  // 验证码
  code?: string;
}

export interface volunteerUserInfo {
  // 加密字符串
  encryptedData?: string;
  // 从微信获取的iv
  iv?: string;
}